Move SelectPage to input module
